fix irq problem
