+++ lab3/include/irq.h
@@ -1,4 +1,6 @@
 #ifndef	_IRQ_H
 #define	_IRQ_H
-void handle_irq(void);
+void handle_irq();
+void handle_sys_timer_irq();
+void handle_core_timer_irq(); 
 #endif
