+++ lab3/include/timer.h
@@ -1,7 +1,7 @@
 #ifndef	_TIMER_H
 #define	_TIMER_H
 
-void sys_timer_init ( void );
-void handle_sys_timer_irq ( void );
-
+void enable_timer_controller();
+void disable_timer_controller();
+void sys_timer_init ();
 #endif  /*_TIMER_H */
