+++ lab3/src/sys.S
@@ -14,3 +14,12 @@ sync_call_uart:
     svc #3
     ret
 
+.globl sync_call_dintp
+sync_call_dintp:
+    svc #4
+    ret
+
+.globl sync_call_eintp
+sync_call_eintp:
+    svc #5
+    ret
